requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
altgraph==0.17.4
annotated-types==0.6.0
anyio==3.7.1
appdirs==1.4.4
astor==0.8.1
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
bcrypt==4.0.1
blessed==1.20.0
boltons==21.0.0
bracex==2.4
certifi==2023.7.22
charset-normalizer==3.3.0
Chroma==0.2.0
chroma-hnswlib==0.7.3
chromadb==0.4.14
click==8.1.7
click-option-group==0.5.6
colorama==0.4.6
coloredlogs==15.0.1
defusedxml==0.7.1
face==22.0.0
fastapi==0.104.0
filelock==3.12.4
flatbuffers==23.5.26
frozenlist==1.4.0
fsspec==2023.9.2
git-python==1.0.3
gitdb==4.0.10
GitPython==3.1.40
glom==22.1.0
grpcio==1.59.0
h11==0.14.0
httptools==0.6.1
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
inquirer==3.1.3
Jinja2==3.1.2
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
litellm==0.1.824
macholib==1.16.3
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
numpy==1.26.1
onnxruntime==1.16.1
ooba==0.0.21
open-interpreter==0.1.7
openai==0.28.1
overrides==7.4.0
packaging==23.2
peewee==3.17.0
posthog==3.0.2
protobuf==4.24.4
pulsar-client==3.3.0
pydantic==2.4.2
pydantic_core==2.10.1
Pygments==2.16.1
pyinstaller==6.1.0
pyinstaller-hooks-contrib==2023.10
PyPika==0.48.9
python-dateutil==2.8.2
python-dotenv==1.0.0
python-editor==1.0.4
python-lsp-jsonrpc==1.0.0
PyYAML==6.0.1
readchar==4.0.5
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
rich==13.6.0
rpds-py==0.10.6
ruamel.yaml==0.17.35
ruamel.yaml.clib==0.2.8
semgrep==1.45.0
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
starlette==0.27.0
sympy==1.12
termcolor==2.3.0
tiktoken==0.4.0
tokenizers==0.14.1
tokentrim==0.1.10
tomli==2.0.1
tqdm==4.66.1
typer==0.9.0
typing_extensions==4.8.0
ujson==5.8.0
urllib3==1.26.18
uvicorn==0.23.2
uvloop==0.18.0
watchfiles==0.21.0
wcmatch==8.5
wcwidth==0.2.8
websockets==11.0.3
wget==3.2
yarl==1.9.2
yaspin==3.0.1
zipp==3.17.0
